+++ java-backend/src/main/java/com/ajmservices/javabackend/repository/CommentaireRepository.java
@@ -3,7 +3,7 @@ package com.ajmservices.javabackend.repository;
 import org.springframework.data.repository.CrudRepository;
 import com.ajmservices.javabackend.model.Commentaire;
 
-public interface CommentaireRepository extends CrudRepository<Commentaire, Integer> {
+public interface CommentaireRepository extends CrudRepository<Commentaire, Long> {
 
     
 }
