+++ java-backend/src/main/java/com/ajmservices/javabackend/repository/EtatRessourceRepository.java
@@ -3,6 +3,6 @@ package com.ajmservices.javabackend.repository;
 import org.springframework.data.repository.CrudRepository;
 import com.ajmservices.javabackend.model.EtatRessource;
 
-public interface EtatRessourceRepository extends CrudRepository<EtatRessource, Integer> {
+public interface EtatRessourceRepository extends CrudRepository<EtatRessource, Long> {
 
 }
