+++ java-backend/src/main/java/com/ajmservices/javabackend/repository/MessageUtilisateurRepository.java
@@ -3,6 +3,6 @@ package com.ajmservices.javabackend.repository;
 import org.springframework.data.repository.CrudRepository;
 import com.ajmservices.javabackend.model.MessageUtilisateur;
 
-public interface MessageUtilisateurRepository extends CrudRepository<MessageUtilisateur, Integer> {
+public interface MessageUtilisateurRepository extends CrudRepository<MessageUtilisateur, Long> {
     
 }
