+++ java-backend/src/main/java/com/ajmservices/javabackend/repository/RessourceRepository.java
@@ -3,6 +3,6 @@ package com.ajmservices.javabackend.repository;
 import org.springframework.data.repository.CrudRepository;
 import com.ajmservices.javabackend.model.Ressource;
 
-public interface RessourceRepository extends CrudRepository<Ressource, Integer> {
+public interface RessourceRepository extends CrudRepository<Ressource, Long> {
     
 }
