+++ java-backend/src/main/java/com/ajmservices/javabackend/repository/TypeRelationRepository.java
@@ -3,6 +3,6 @@ package com.ajmservices.javabackend.repository;
 import org.springframework.data.repository.CrudRepository;
 import com.ajmservices.javabackend.model.TypeRelation;
 
-public interface TypeRelationRepository extends CrudRepository<TypeRelation, Integer> {
+public interface TypeRelationRepository extends CrudRepository<TypeRelation, Long> {
     
 }
