+++ java-backend/src/main/java/com/ajmservices/javabackend/repository/TypeRessourceRepository.java
@@ -6,6 +6,6 @@ import com.ajmservices.javabackend.model.TypeRessource;
 // This will be AUTO IMPLEMENTED by Spring into a Bean called userRepository
 // CRUD refers Create, Read, Update, Delete
 
-public interface TypeRessourceRepository extends CrudRepository<TypeRessource, Integer> {
+public interface TypeRessourceRepository extends CrudRepository<TypeRessource, Long> {
 
 }
