+++ java-backend/src/main/java/com/ajmservices/javabackend/repository/UtilisateurRepository.java
@@ -3,6 +3,6 @@ package com.ajmservices.javabackend.repository;
 import org.springframework.data.repository.CrudRepository;
 import com.ajmservices.javabackend.model.Utilisateur;
 
-public interface UtilisateurRepository extends CrudRepository<Utilisateur, Integer>{
+public interface UtilisateurRepository extends CrudRepository<Utilisateur, Long>{
     
 }
